backend/backend-cloud-auth/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Alexander Morales Panitz <alexander@moralespanitz.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
markdown = "^3.6"
django-filter = "^24.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pyjwt = "^2.8.0"
requests = "^2.32.3"
cryptography = "^42.0.8"
pytz = "^2024.1"
mysqlclient = "^2.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
